feat: Add MessageBanner, SuccessModal, and TimeSlotSelector components for booking notifications and time selection
feat: Implement BookingCard and BookingsList components for managing student bookings

feat: Create CancelModal and UpdateModal components for booking cancellation and updates

feat: Add FilterControls component for filtering and sorting bookings

style: Enhance UI with dark mode support and improved accessibility features
