+++ frontend/ta_connect/src/pages/student/ManageBookingsPage.jsx
@@ -5,6 +5,11 @@ import { useGlobalLoading } from '../../contexts/GlobalLoadingContext';
 import { useLocation } from 'react-router-dom';
 import StudentNavbar from '../../components/student/studentNavbar';
 import Footer from '../../components/General/Footer';
+import FilterControls from '../../components/student/manage-bookings/FilterControls';
+import BookingsList from '../../components/student/manage-bookings/BookingsList';
+import CancelModal from '../../components/student/manage-bookings/CancelModal';
+import UpdateModal from '../../components/student/manage-bookings/UpdateModal';
+import MessageBanner from '../../components/student/booking/MessageBanner';
 import axios from 'axios';
 import { useQuery } from '@tanstack/react-query';
 import { useCancelInstructorBooking, useUpdateBooking } from '../../hooks/useApi';
@@ -395,173 +400,38 @@ export default function ManageBookingsPage() {
             </div>
 
             {/* Filter and Sort Controls */}
-            <div className={`${isDark ? 'bg-gray-800' : 'bg-white'} p-4 sm:p-6 rounded-xl shadow-lg mb-4 sm:mb-6`}>
-              <div className="flex flex-col gap-3 sm:gap-4">
-                {/* Top Row: Filters */}
-                <div className="flex flex-row gap-3 sm:gap-4">
-                  {/* Status Filter */}
-                  <div className="flex-1 min-w-[150px]">
-                    <label className={`block text-xs sm:text-sm font-medium ${isDark ? 'text-gray-300' : 'text-gray-700'} mb-1.5 sm:mb-2`}>
-                      {t.filters.status}
-                    </label>
-                    <select
-                      value={filterStatus}
-                      onChange={(e) => setFilterStatus(e.target.value)}
-                      className={`w-full px-4 py-2 rounded-lg border-2 ${
-                        isDark 
-                          ? 'bg-gray-700 border-gray-600 text-white focus:border-[#366c6b]' 
-                          : 'bg-white border-gray-300 text-gray-900 focus:border-[#366c6b]'
-                      } focus:outline-none transition-colors`}
-                    >
-                      <option value="all">{t.filters.allBookings}</option>
-                      <option value="active">{t.filters.activeOnly}</option>
-                      <option value="cancelled">{t.filters.cancelledOnly}</option>
-                      <option value="completed">{t.filters.completedOnly}</option>
-                    </select>
-                  </div>
-
-                  {/* Course Filter */}
-                  <div className="flex-1 min-w-[150px]">
-                    <label className={`block text-xs sm:text-sm font-medium ${isDark ? 'text-gray-300' : 'text-gray-700'} mb-1.5 sm:mb-2`}>
-                      {t.filters.course}
-                    </label>
-                    <select
-                      value={filterCourse}
-                      onChange={(e) => setFilterCourse(e.target.value)}
-                      className={`w-full px-4 py-2 rounded-lg border-2 ${
-                        isDark 
-                          ? 'bg-gray-700 border-gray-600 text-white focus:border-[#366c6b]' 
-                          : 'bg-white border-gray-300 text-gray-900 focus:border-[#366c6b]'
-                      } focus:outline-none transition-colors`}
-                    >
-                      <option value="all">{t.filters.allCourses}</option>
-                      {uniqueCourses.map(course => (
-                        <option key={course} value={course}>{course}</option>
-                      ))}
-                    </select>
-                  </div>
-
-                  {/* Sort By */}
-                  <div className="flex-1 min-w-[150px]">
-                    <label className={`block text-xs sm:text-sm font-medium ${isDark ? 'text-gray-300' : 'text-gray-700'} mb-1.5 sm:mb-2`}>
-                      {t.filters.sortBy}
-                    </label>
-                    <select
-                      value={sortBy}
-                      onChange={(e) => setSortBy(e.target.value)}
-                      className={`w-full px-4 py-2 rounded-lg border-2 ${
-                        isDark 
-                          ? 'bg-gray-700 border-gray-600 text-white focus:border-[#366c6b]' 
-                          : 'bg-white border-gray-300 text-gray-900 focus:border-[#366c6b]'
-                      } focus:outline-none transition-colors`}
-                    >
-                      <option value="date">{t.filters.byDate}</option>
-                      <option value="course">{t.filters.byCourse}</option>
-                    </select>
-                  </div>
-                </div>
-
-                {/* Date Range */}
-                <div className="grid grid-cols-1 sm:grid-cols-2 gap-3 sm:gap-4">
-                  <div>
-                    <label className={`block text-xs sm:text-sm font-medium ${isDark ? 'text-gray-300' : 'text-gray-700'} mb-1.5 sm:mb-2`}>
-                      {t.filters.from}
-                    </label>
-                    <input
-                      type="date"
-                      value={dateRange.start}
-                      onChange={(e) => setDateRange(prev => ({ ...prev, start: e.target.value }))}
-                      className={`w-full px-4 py-2 rounded-lg border-2 ${
-                        isDark
-                          ? 'bg-gray-700 border-gray-600 text-white focus:border-[#366c6b]'
-                          : 'bg-white border-gray-300 text-gray-900 focus:border-[#366c6b]'
-                      } focus:outline-none transition-colors`}
-                    />
-                  </div>
-                  <div>
-                    <label className={`block text-xs sm:text-sm font-medium ${isDark ? 'text-gray-300' : 'text-gray-700'} mb-1.5 sm:mb-2`}>
-                      {t.filters.to}
-                    </label>
-                    <input
-                      type="date"
-                      value={dateRange.end}
-                      onChange={(e) => setDateRange(prev => ({ ...prev, end: e.target.value }))}
-                      className={`w-full px-4 py-2 rounded-lg border-2 ${
-                        isDark
-                          ? 'bg-gray-700 border-gray-600 text-white focus:border-[#366c6b]'
-                          : 'bg-white border-gray-300 text-gray-900 focus:border-[#366c6b]'
-                      } focus:outline-none transition-colors`}
-                    />
-                  </div>
-                </div>
-
-                {/* Bottom Row: Action Buttons */}
-                <div className="flex flex-row gap-2 pt-2 sm:pt-3 border-t border-gray-600/30">
-                  <button
-                    onClick={handleClearFilters}
-                    className={`w-auto px-4 py-2 rounded-lg border transition-all ${
-                      isDark
-                        ? 'border-gray-500 text-gray-300 hover:bg-gray-600'
-                        : 'border-gray-300 text-gray-700 hover:bg-gray-100'
-                    }`}
-                  >
-                    {t.filters.clear}
-                  </button>
-                  {bookings.some(b => b.is_cancelled) && (
-                    <button
-                      onClick={handleClearCancelled}
-                      className={`w-auto px-4 py-2 rounded-lg font-medium text-sm transition-all duration-300 flex items-center justify-center gap-2 ${
-                        isDark 
-                          ? 'bg-red-900/30 text-red-300 hover:bg-red-900/50 border-2 border-red-700 hover:scale-105' 
-                          : 'bg-red-50 text-red-700 hover:bg-red-100 border-2 border-red-300 hover:scale-105'
-                      } shadow-md hover:shadow-lg`}
-                    >
-                      <svg className="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                        <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M19 7l-.867 12.142A2 2 0 0116.138 21H7.862a2 2 0 01-1.995-1.858L5 7m5 4v6m4-6v6m1-10V4a1 1 0 00-1-1h-4a1 1 0 00-1 1v3M4 7h16" />
-                      </svg>
-                      {t.buttons.deleteAllCancelled}
-                    </button>
-                  )}
-                  {bookings.some(b => b.is_completed) && (
-                    <button
-                      onClick={handleClearCompleted}
-                      className={`w-auto px-4 py-2 rounded-lg font-medium text-sm transition-all duration-300 flex items-center justify-center gap-2 ${
-                        isDark 
-                          ? 'bg-green-900/30 text-green-300 hover:bg-green-900/50 border-2 border-green-700 hover:scale-105' 
-                          : 'bg-green-50 text-green-700 hover:bg-green-100 border-2 border-green-300 hover:scale-105'
-                      } shadow-md hover:shadow-lg`}
-                    >
-                      <svg className="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                        <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M19 7l-.867 12.142A2 2 0 0116.138 21H7.862a2 2 0 01-1.995-1.858L5 7m5 4v6m4-6v6m1-10V4a1 1 0 00-1-1h-4a1 1 0 00-1 1v3M4 7h16" />
-                      </svg>
-                      {t.buttons.deleteAllCompleted || 'Clear Completed'}
-                    </button>
-                  )}
-                </div>
-              </div>
-            </div>
+            <FilterControls
+              filterStatus={filterStatus}
+              setFilterStatus={setFilterStatus}
+              filterCourse={filterCourse}
+              setFilterCourse={setFilterCourse}
+              sortBy={sortBy}
+              setSortBy={setSortBy}
+              dateRange={dateRange}
+              setDateRange={setDateRange}
+              uniqueCourses={uniqueCourses}
+              hasCancelled={bookings.some(b => b.is_cancelled)}
+              hasCompleted={bookings.some(b => b.is_completed)}
+              onClearFilters={handleClearFilters}
+              onClearCancelled={handleClearCancelled}
+              onClearCompleted={handleClearCompleted}
+            />
 
             {/* Success/Error Messages */}
             {success && (
-              <div className={`mb-6 p-4 ${isDark ? 'bg-green-900/30 border-green-600' : 'bg-green-50 border-green-300'} border-2 rounded-xl`}>
-                <div className="flex items-start">
-                  <svg className={`w-5 h-5 mr-3 mt-0.5 ${isDark ? 'text-green-400' : 'text-green-600'}`} fill="currentColor" viewBox="0 0 20 20">
-                    <path fillRule="evenodd" d="M10 18a8 8 0 100-16 8 8 0 000 16zm3.707-9.293a1 1 0 00-1.414-1.414L9 10.586 7.707 9.293a1 1 0 00-1.414 1.414l2 2a1 1 0 001.414 0l4-4z" clipRule="evenodd" />
-                  </svg>
-                  <span className={`${isDark ? 'text-green-200' : 'text-green-700'} font-medium`}>{success}</span>
-                </div>
-              </div>
+              <MessageBanner
+                type="success"
+                message={success}
+                onClose={() => setSuccess('')}
+              />
             )}
 
             {error && (
-              <div className={`mb-6 p-4 ${isDark ? 'bg-red-900/30 border-red-600' : 'bg-red-50 border-red-300'} border-2 rounded-xl`}>
-                <div className="flex items-start">
-                  <svg className={`w-5 h-5 mr-3 mt-0.5 ${isDark ? 'text-red-400' : 'text-red-600'}`} fill="currentColor" viewBox="0 0 20 20">
-                    <path fillRule="evenodd" d="M10 18a8 8 0 100-16 8 8 0 000 16zM8.707 7.293a1 1 0 00-1.414 1.414L8.586 10l-1.293 1.293a1 1 0 101.414 1.414L10 11.414l1.293 1.293a1 1 0 001.414-1.414L11.414 10l1.293-1.293a1 1 0 00-1.414-1.414L10 8.586 8.707 7.293z" clipRule="evenodd" />
-                  </svg>
-                  <span className={`${isDark ? 'text-red-200' : 'text-red-700'} font-medium`}>{error}</span>
-                </div>
-              </div>
+              <MessageBanner
+                type="error"
+                message={error}
+                onClose={() => setError('')}
+              />
             )}
 
             {/* Loading State */}
@@ -573,196 +443,41 @@ export default function ManageBookingsPage() {
               <>
                 {/* Active Bookings */}
                 {filterStatus !== 'cancelled' && filterStatus !== 'completed' && (
-                  <div className={`${isDark ? 'bg-gray-800' : 'bg-white'} rounded-xl shadow-lg mb-4 sm:mb-6 p-4 sm:p-6`}>
-                    <h2 className={`text-xl sm:text-2xl font-bold ${isDark ? 'text-white' : 'text-gray-900'} mb-4 sm:mb-6`}>
-                      {t.sections.activeBookings} ({activeBookings.length})
-                    </h2>
-                  
-                  {activeBookings.length === 0 ? (
-                    <div className="text-center py-12">
-                      <svg className={`mx-auto h-12 w-12 ${isDark ? 'text-gray-600' : 'text-gray-400'}`} fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                        <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M8 7V3m8 4V3m-9 8h10M5 21h14a2 2 0 002-2V7a2 2 0 00-2-2H5a2 2 0 00-2 2v12a2 2 0 002 2z" />
-                      </svg>
-                      <p className={`mt-4 text-lg ${isDark ? 'text-gray-400' : 'text-gray-600'}`}>
-                        {t.messages.noActiveBookings}
-                      </p>
-                    </div>
-                  ) : (
-                    <div className="grid grid-cols-1 lg:grid-cols-2 gap-3 sm:gap-4">
-                      {activeBookings.map((booking) => (
-                        <div
-                          key={booking.id}
-                          className={`${isDark ? 'bg-gray-700 border-gray-600' : 'bg-gradient-to-br from-[#eaf6f6] to-white border-[#366c6b]/20'} border-2 rounded-xl p-4 sm:p-6 transition-all duration-300 hover:shadow-lg`}
-                        >
-                          <div className="flex justify-between items-start mb-3 sm:mb-4">
-                            <div>
-                              <h3 className={`text-lg sm:text-xl font-bold ${isDark ? 'text-white' : 'text-gray-900'}`}>
-                                {booking.course_name}
-                              </h3>
-                              {booking.section && (
-                                <p className={`text-xs sm:text-sm ${isDark ? 'text-gray-400' : 'text-gray-600'}`}>
-                                  {t.bookingCard.section} {booking.section}
-                                </p>
-                              )}
-                            </div>
-                          </div>
-                          
-                          <div className={`space-y-1.5 sm:space-y-2 mb-3 sm:mb-4 text-sm sm:text-base ${isDark ? 'text-gray-300' : 'text-gray-700'}`}>
-                            <p className="flex items-center gap-2">
-                              <span>👨‍🏫</span>
-                              <span dir="ltr">{booking.instructor?.full_name || 'Instructor'}</span>
-                            </p>
-                            <p className="flex items-center gap-2">
-                              <span>📅</span>
-                              <span dir="ltr">{formatDate(booking.date)}</span>
-                            </p>
-                            <p className="flex items-center gap-2">
-                              <span>🕐</span>
-                              <span dir="ltr">{formatTime(booking.start_time)} - {formatTime(booking.end_time)}</span>
-                            </p>
-                            {booking.room && (
-                              <p className="flex items-center gap-2">
-                                <span>📍</span>
-                                <span dir="ltr">{booking.room}</span>
-                              </p>
-                            )}
-                          </div>
-
-                          <div className="flex flex-col xs:flex-row gap-2">
-                            <button
-                              onClick={() => handleUpdateClick(booking)}
-                              className="flex-1 px-4 py-2 sm:py-2.5 bg-gradient-to-r from-[#366c6b] to-[#1a3535] text-white rounded-lg hover:shadow-lg transition-all duration-300 hover:scale-105 text-xs sm:text-sm font-medium"
-                            >
-                              {t.buttons.update}
-                            </button>
-                            <button
-                              onClick={() => handleCancelClick(booking)}
-                              className={`flex-1 px-4 py-2 sm:py-2.5 ${isDark ? 'bg-red-900/30 text-red-300 hover:bg-red-900/50' : 'bg-red-100 text-red-700 hover:bg-red-200'} rounded-lg transition-all duration-300 text-xs sm:text-sm font-medium`}
-                            >
-                              {t.buttons.cancel}
-                            </button>
-                          </div>
-                        </div>
-                      ))}
-                    </div>
-                  )}
-                  </div>
+                  <BookingsList
+                    title={t.sections.activeBookings}
+                    bookings={activeBookings}
+                    status="active"
+                    emptyMessage={t.messages.noActiveBookings}
+                    formatDate={formatDate}
+                    formatTime={formatTime}
+                    onUpdate={handleUpdateClick}
+                    onCancel={handleCancelClick}
+                  />
                 )}
 
                 {/* Cancelled Bookings */}
                 {filterStatus !== 'active' && filterStatus !== 'completed' && (
-                  <div className={`${isDark ? 'bg-gray-800' : 'bg-white'} rounded-xl shadow-lg mb-4 sm:mb-6 p-4 sm:p-6`}>
-                    <h2 className={`text-xl sm:text-2xl font-bold ${isDark ? 'text-white' : 'text-gray-900'} mb-4 sm:mb-6`}>
-                      {t.sections.cancelledBookings} ({cancelledBookings.length})
-                    </h2>
-                    
-                  {cancelledBookings.length === 0 ? (
-                    <div className="text-center py-12">
-                      <svg className={`mx-auto h-12 w-12 ${isDark ? 'text-gray-600' : 'text-gray-400'}`} fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                        <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M6 18L18 6M6 6l12 12" />
-                      </svg>
-                      <p className={`mt-4 text-lg ${isDark ? 'text-gray-400' : 'text-gray-600'}`}>
-                        {t.messages.noCancelledBookings}
-                      </p>
-                    </div>
-                  ) : (
-                    <div className="grid grid-cols-1 lg:grid-cols-2 gap-3 sm:gap-4">
-                      {cancelledBookings.map((booking) => (
-                        <div
-                          key={booking.id}
-                          className={`${isDark ? 'bg-gray-700/50 border-gray-600' : 'bg-gray-50 border-gray-300'} border-2 rounded-xl p-4 sm:p-6 opacity-75`}
-                        >
-                          <div className="flex justify-between items-start mb-3 sm:mb-4">
-                            <div>
-                              <h3 className={`text-lg sm:text-xl font-bold ${isDark ? 'text-white' : 'text-gray-900'}`}>
-                                {booking.course_name}
-                              </h3>
-                              <span className={`inline-block mt-1 px-2 py-1 text-xs font-semibold rounded ${isDark ? 'bg-red-900/30 text-red-300' : 'bg-red-100 text-red-700'}`}>
-                                {t.status.cancelled}
-                              </span>
-                            </div>
-                          </div>
-                          
-                          <div className={`space-y-1.5 sm:space-y-2 text-sm sm:text-base ${isDark ? 'text-gray-400' : 'text-gray-600'}`}>
-                            <p className="flex items-center gap-2">
-                              <span>👨‍🏫</span>
-                              <span dir="ltr">{booking.instructor?.full_name || 'Instructor'}</span>
-                            </p>
-                            <p className="flex items-center gap-2">
-                              <span>📅</span>
-                              <span dir="ltr">{formatDate(booking.date)}</span>
-                            </p>
-                            <p className="flex items-center gap-2">
-                              <span>🕐</span>
-                              <span dir="ltr">{formatTime(booking.start_time)} - {formatTime(booking.end_time)}</span>
-                            </p>
-                          </div>
-                        </div>
-                      ))}
-                    </div>
-                  )}
-                  </div>
+                  <BookingsList
+                    title={t.sections.cancelledBookings}
+                    bookings={cancelledBookings}
+                    status="cancelled"
+                    emptyMessage={t.messages.noCancelledBookings}
+                    formatDate={formatDate}
+                    formatTime={formatTime}
+                  />
                 )}
 
                 {/* Completed Bookings */}
                 {filterStatus !== 'active' && filterStatus !== 'cancelled' && (
-                  <div className={`${isDark ? 'bg-gray-800' : 'bg-white'} rounded-xl shadow-lg p-4 sm:p-6`}>
-                    <h2 className={`text-xl sm:text-2xl font-bold ${isDark ? 'text-white' : 'text-gray-900'} mb-4 sm:mb-6`}>
-                      {t.sections.completedBookings} ({completedBookings.length})
-                    </h2>
-                    
-                  {completedBookings.length === 0 ? (
-                    <div className="text-center py-12">
-                      <svg className={`mx-auto h-12 w-12 ${isDark ? 'text-gray-600' : 'text-gray-400'}`} fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                        <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M9 12l2 2 4-4m6 2a9 9 0 11-18 0 9 9 0 0118 0z" />
-                      </svg>
-                      <p className={`mt-4 text-lg ${isDark ? 'text-gray-400' : 'text-gray-600'}`}>
-                        {t.messages.noCompletedBookings}
-                      </p>
-                    </div>
-                  ) : (
-                    <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-3 sm:gap-4">
-                      {completedBookings.map((booking) => (
-                        <div
-                          key={booking.id}
-                          className={`${isDark ? 'bg-gradient-to-br from-gray-700 to-gray-750 border-green-600/30' : 'bg-gradient-to-br from-green-50 to-white border-green-300'} border-2 rounded-xl p-4 transition-all duration-300 hover:shadow-lg`}
-                        >
-                          <div className="flex justify-between items-start mb-3">
-                            <div>
-                              <h3 className={`text-base sm:text-lg font-bold ${isDark ? 'text-white' : 'text-gray-900'}`}>
-                                {booking.course_name}
-                              </h3>
-                              <span className={`inline-block mt-1 px-2 py-0.5 text-xs font-semibold rounded ${isDark ? 'bg-green-900/30 text-green-300' : 'bg-green-100 text-green-700'}`}>
-                                {t.status.completed}
-                              </span>
-                            </div>
-                          </div>
-                          
-                          <div className={`space-y-1.5 text-sm ${isDark ? 'text-gray-300' : 'text-gray-700'}`}>
-                            <p className="flex items-center gap-2">
-                              <span>👨‍🏫</span>
-                              <span className="truncate" dir="ltr">{booking.instructor?.full_name || 'Instructor'}</span>
-                            </p>
-                            <p className="flex items-center gap-2">
-                              <span>📅</span>
-                              <span dir="ltr">{formatDate(booking.date)}</span>
-                            </p>
-                            <p className="flex items-center gap-2">
-                              <span>🕐</span>
-                              <span dir="ltr">{formatTime(booking.start_time)} - {formatTime(booking.end_time)}</span>
-                            </p>
-                            {booking.room && (
-                              <p className="flex items-center gap-2">
-                                <span>📍</span>
-                                <span className="truncate" dir="ltr">{booking.room}</span>
-                              </p>
-                            )}
-                          </div>
-                        </div>
-                      ))}
-                    </div>
-                  )}
-                  </div>
+                  <BookingsList
+                    title={t.sections.completedBookings}
+                    bookings={completedBookings}
+                    status="completed"
+                    emptyMessage={t.messages.noCompletedBookings}
+                    formatDate={formatDate}
+                    formatTime={formatTime}
+                    gridCols="grid-cols-1 md:grid-cols-2 lg:grid-cols-3"
+                  />
                 )}
               </>
             )}
@@ -773,160 +488,32 @@ export default function ManageBookingsPage() {
       <Footer />
 
       {/* Cancel Confirmation Modal */}
-      {showCancelModal && selectedBooking && (
-        <div className="fixed inset-0 bg-black bg-opacity-50 flex items-center justify-center z-50 p-3 sm:p-4">
-          <div className={`${isDark ? 'bg-gray-800' : 'bg-white'} rounded-xl shadow-2xl max-w-md w-full p-4 sm:p-6`}>
-            <h3 className={`text-xl sm:text-2xl font-bold ${isDark ? 'text-white' : 'text-gray-900'} mb-3 sm:mb-4`}>
-              {t.modals.cancelTitle}
-            </h3>
-            <p className={`text-sm sm:text-base ${isDark ? 'text-gray-300' : 'text-gray-600'} mb-4 sm:mb-6`}>
-              {t.modals.cancelMessage}
-            </p>
-            
-            <div className={`${isDark ? 'bg-gray-700' : 'bg-gray-50'} p-3 sm:p-4 rounded-lg mb-4`}>
-              <p className={`font-semibold ${isDark ? 'text-white' : 'text-gray-900'}`}>
-                {selectedBooking.course_name}
-              </p>
-              <p className={`text-sm ${isDark ? 'text-gray-300' : 'text-gray-600'} mt-2`} dir="ltr">
-                📅 {formatDate(selectedBooking.date)}
-              </p>
-              <p className={`text-sm ${isDark ? 'text-gray-300' : 'text-gray-600'}`} dir="ltr">
-                🕐 {formatTime(selectedBooking.start_time)} - {formatTime(selectedBooking.end_time)}
-              </p>
-            </div>
-
-            <div className="flex gap-3">
-              <button
-                onClick={() => setShowCancelModal(false)}
-                disabled={cancelLoading}
-                className={`flex-1 px-4 py-2 ${isDark ? 'bg-gray-700 text-white hover:bg-gray-600' : 'bg-gray-200 text-gray-900 hover:bg-gray-300'} rounded-lg transition-all duration-300 font-medium`}
-              >
-                {t.modals.keepBooking}
-              </button>
-              <button
-                onClick={handleCancelBooking}
-                disabled={cancelLoading}
-                className="flex-1 px-4 py-2 bg-red-600 text-white rounded-lg hover:bg-red-700 transition-all duration-300 font-medium disabled:opacity-50 flex items-center justify-center"
-              >
-                {cancelLoading ? (
-                  <svg className="animate-spin h-5 w-5" fill="none" viewBox="0 0 24 24">
-                    <circle className="opacity-25" cx="12" cy="12" r="10" stroke="currentColor" strokeWidth="4"></circle>
-                    <path className="opacity-75" fill="currentColor" d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"></path>
-                  </svg>
-                ) : (
-                  t.modals.confirmCancel
-                )}
-              </button>
-            </div>
-          </div>
-        </div>
-      )}
+      <CancelModal
+        show={showCancelModal}
+        booking={selectedBooking}
+        formatDate={formatDate}
+        formatTime={formatTime}
+        onConfirm={handleCancelBooking}
+        onClose={() => setShowCancelModal(false)}
+        isLoading={cancelLoading}
+      />
 
       {/* Update Booking Modal */}
-      {showUpdateModal && selectedBooking && (
-        <div className="fixed inset-0 bg-black bg-opacity-50 flex items-center justify-center z-50 p-3 sm:p-4 overflow-y-auto">
-          <div className={`${isDark ? 'bg-gray-800' : 'bg-white'} rounded-xl shadow-2xl max-w-2xl w-full p-4 sm:p-6 my-4 sm:my-8`}>
-            <h3 className={`text-xl sm:text-2xl font-bold ${isDark ? 'text-white' : 'text-gray-900'} mb-3 sm:mb-4`}>
-              {t.modals.updateTitle}
-            </h3>
-            
-            <div className={`${isDark ? 'bg-gray-700' : 'bg-gray-50'} p-4 rounded-lg mb-6`}>
-              <p className={`font-semibold ${isDark ? 'text-white' : 'text-gray-900'}`}>
-                {t.modals.currentBooking}
-              </p>
-              <p className={`text-sm ${isDark ? 'text-gray-300' : 'text-gray-600'} mt-2`} dir="ltr">
-                📅 {formatDate(selectedBooking.date)}
-              </p>
-              <p className={`text-sm ${isDark ? 'text-gray-300' : 'text-gray-600'}`} dir="ltr">
-                🕐 {formatTime(selectedBooking.start_time)} - {formatTime(selectedBooking.end_time)}
-              </p>
-            </div>
-
-            <div className="mb-6">
-              <h4 className={`font-semibold ${isDark ? 'text-white' : 'text-gray-900'} mb-3`}>
-                {t.modals.selectNewDate}
-              </h4>
-              <div className="max-h-48 overflow-y-auto space-y-2">
-                {availableDates.map((date, index) => {
-                  // Use local timezone date formatting to avoid UTC conversion issues
-                  const yyyy = date.getFullYear();
-                  const mm = String(date.getMonth() + 1).padStart(2, '0');
-                  const dd = String(date.getDate()).padStart(2, '0');
-                  const dateStr = `${yyyy}-${mm}-${dd}`;
-                  return (
-                    <button
-                      key={index}
-                      onClick={() => handleDateSelect(dateStr)}
-                      className={`w-full p-3 rounded-lg border-2 text-left transition-all ${
-                        newDate === dateStr
-                          ? isDark ? 'border-[#366c6b] bg-[#366c6b]/20' : 'border-[#366c6b] bg-[#366c6b]/10'
-                          : isDark ? 'border-gray-600 hover:border-gray-500 bg-gray-700' : 'border-gray-200 hover:border-gray-300 bg-white'
-                      }`}
-                    >
-                      <p className={`font-semibold ${isDark ? 'text-white' : 'text-gray-900'}`}>
-                        {formatDate(date)}
-                      </p>
-                    </button>
-                  );
-                })}
-              </div>
-            </div>
-
-            {newDate && (
-              <div className="mb-6">
-                <h4 className={`font-semibold ${isDark ? 'text-white' : 'text-gray-900'} mb-3`}>
-                  {t.modals.selectNewTime}
-                </h4>
-                {availableTimes.length === 0 ? (
-                  <p className={`text-center py-4 ${isDark ? 'text-gray-400' : 'text-gray-600'}`}>
-                    {t.modals.noAvailableTimes}
-                  </p>
-                ) : (
-                  <div className="grid grid-cols-2 sm:grid-cols-3 gap-2 max-h-48 overflow-y-auto">
-                    {availableTimes.map((time, index) => (
-                      <button
-                        key={index}
-                        onClick={() => setNewTime(time)}
-                        className={`p-2 sm:p-2.5 rounded-lg border-2 text-center transition-all font-medium text-xs sm:text-sm ${
-                          newTime === time
-                            ? isDark ? 'border-[#366c6b] bg-[#366c6b]/20 text-white' : 'border-[#366c6b] bg-[#366c6b]/10 text-[#366c6b]'
-                            : isDark ? 'border-gray-600 hover:border-gray-500 bg-gray-700 text-gray-300' : 'border-gray-200 hover:border-gray-300 bg-white text-gray-700'
-                        }`}
-                      >
-                        {formatTime(time)}
-                      </button>
-                    ))}
-                  </div>
-                )}
-              </div>
-            )}
-
-            <div className="flex gap-3">
-              <button
-                onClick={() => setShowUpdateModal(false)}
-                disabled={isUpdating}
-                className={`flex-1 px-4 py-2 ${isDark ? 'bg-gray-700 text-white hover:bg-gray-600' : 'bg-gray-200 text-gray-900 hover:bg-gray-300'} rounded-lg transition-all duration-300 font-medium`}
-              >
-                {t.modals.cancelButton}
-              </button>
-              <button
-                onClick={handleUpdateBooking}
-                disabled={isUpdating || !newDate || !newTime}
-                className="flex-1 px-4 py-2 bg-gradient-to-r from-[#366c6b] to-[#1a3535] text-white rounded-lg hover:shadow-lg transition-all duration-300 font-medium disabled:opacity-50 flex items-center justify-center"
-              >
-                {isUpdating ? (
-                  <svg className="animate-spin h-5 w-5" fill="none" viewBox="0 0 24 24">
-                    <circle className="opacity-25" cx="12" cy="12" r="10" stroke="currentColor" strokeWidth="4"></circle>
-                    <path className="opacity-75" fill="currentColor" d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"></path>
-                  </svg>
-                ) : (
-                  t.modals.confirmUpdate
-                )}
-              </button>
-            </div>
-          </div>
-        </div>
-      )}
+      <UpdateModal
+        show={showUpdateModal}
+        booking={selectedBooking}
+        availableDates={availableDates}
+        availableTimes={availableTimes}
+        newDate={newDate}
+        newTime={newTime}
+        formatDate={formatDate}
+        formatTime={formatTime}
+        onDateSelect={handleDateSelect}
+        onTimeSelect={setNewTime}
+        onConfirm={handleUpdateBooking}
+        onClose={() => setShowUpdateModal(false)}
+        isLoading={isUpdating}
+      />
     </div>
   );
 }
